+++ database/migrations/1761938986254_create_categorias_table.ts
@@ -9,9 +9,9 @@ export default class extends BaseSchema {
       table.string('titulo', 255).notNullable();
       table.string('descricao');
       table.boolean('ativo').defaultTo(true);
-      table.integer('criado_por').unsigned().references('cpf').inTable('usuario').notNullable();
+      table.bigInteger('criado_por').unsigned().references('cpf').inTable('usuario').notNullable();
       table.timestamp('criado_em', { useTz: false }).notNullable().defaultTo(this.now());
-      table.integer('atualizado_por').unsigned().references('cpf').inTable('usuario');
+      table.bigInteger('atualizado_por').unsigned().references('cpf').inTable('usuario');
       table.timestamp('atualizado_em', { useTz: false});
     })
   }
