+++ database/migrations/1761939234375_create_impedimentos_table.ts
@@ -11,9 +11,9 @@ export default class extends BaseSchema {
       table.enum('nivel_energia_fisica', Object.values(Nivel) ,{useNative:true, enumName: 'nivel_enerigia_enum'}).defaultTo(Nivel[1]);
       table.enum('nivel_energia_mental',Object.values(Nivel)).defaultTo(Nivel[1]);
       table.boolean('ativo').defaultTo(true);
-      table.integer('criado_por').unsigned().references('cpf').inTable('usuario').notNullable();
+      table.bigInteger('criado_por').unsigned().references('cpf').inTable('usuario').notNullable();
       table.timestamp('criado_em', { useTz: false }).notNullable().defaultTo(this.now());
-      table.integer('atualizado_por').unsigned().references('cpf').inTable('usuario');
+      table.bigInteger('atualizado_por').unsigned().references('cpf').inTable('usuario');
       table.timestamp('atualizado_em', { useTz: false});
     })
   }
