+++ database/migrations/1761941525630_create_sub_tarefa.ts
@@ -19,9 +19,9 @@ export default class extends BaseSchema {
       table.integer('num_tarefa').unsigned().references('numero').inTable('tarefa').notNullable().onDelete('CASCADE');
       table.integer('num_motivacao').unsigned().references('numero').inTable('motivacao').onDelete('SET NULL');
       table.boolean('ativo').defaultTo(true);
-      table.integer('criado_por').unsigned().references('cpf').inTable('usuario').notNullable();
+      table.bigInteger('criado_por').unsigned().references('cpf').inTable('usuario').notNullable();
       table.timestamp('criado_em', { useTz: false }).notNullable().defaultTo(this.now());
-      table.integer('atualizado_por').unsigned().references('cpf').inTable('usuario');
+      table.bigInteger('atualizado_por').unsigned().references('cpf').inTable('usuario');
       table.timestamp('atualizado_em', { useTz: false});
     })
   }
